update stream detail layout
