+++ packages/www/components/Dashboard/Chart/index.tsx
@@ -34,11 +34,18 @@ const CustomTooltip = ({ active, payload }: any) => {
 
 const Chart = ({ data }) => {
   return (
-    <Box css={{ width: "100%", position: "relative" }}>
+    <Box
+      css={{
+        width: "100%",
+        position: "relative",
+        ".recharts-cartesian-axis-tick": {
+          fontSize: "$2",
+        },
+      }}>
       <Text
+        variant="gray"
+        size="1"
         css={{
-          color: "#666666",
-          fontSize: "12px",
           transform: "rotate(-90deg)",
           position: "absolute",
           left: "-70px",
@@ -47,9 +54,9 @@ const Chart = ({ data }) => {
         kbps (multiplied by 1000)
       </Text>
       <Text
+        variant="gray"
+        size="1"
         css={{
-          color: "#666666",
-          fontSize: "12px",
           position: "absolute",
           bottom: "-30px",
           left: "50px",
@@ -60,8 +67,18 @@ const Chart = ({ data }) => {
         <AreaChart data={data}>
           <defs>
             <linearGradient id="colorKbps" x1="0" y1="0" x2="0" y2="1">
-              <stop offset="5%" stopColor="#6E56CF" stopOpacity={0.8} />
-              <stop offset="95%" stopColor="#6E56CF" stopOpacity={0} />
+              <Box
+                as="stop"
+                offset="5%"
+                stopColor="$violet9"
+                stopOpacity={0.8}
+              />
+              <Box
+                as="stop"
+                offset="95%"
+                stopColor="$violet9"
+                stopOpacity={0}
+              />
             </linearGradient>
           </defs>
           <XAxis dataKey="name" />
